+++ docs/src/changelog.md
@@ -6,6 +6,11 @@
 ### New features
 
 - Added `resample` method for `BinnedWeightedResampling` scheme.
+- Added `AbstractBinnedResampling` abstract type for binnings where the values in each 
+    bin is represented by an uncertain value. `BinnedResampling` and `BinnedWeightedResampling` are subtypes `AbstractBinnedResampling`.
+- Added `AbstractBinnedSummarisedResampling` abstract type for binnings where the values 
+    in each bin are summarised to a single value. `BinnedMeanResampling` and
+     `BinnedMeanWeightedResampling` are subtypes `AbstractBinnedResampling`.
 
 ### Improvements